requirements.txt
python-telegram-bot>=20.0
cloudscraper>=1.2.71
beautifulsoup4>=4.12.0
loguru>=0.7.0
python-dotenv>=1.0.0
